fix with float

+++ finance/functions.py
@@ -9,10 +9,10 @@ def last_maximum_occurrence_index(x):
 def rescale_by_part(x, first, minimum, maximum):
     if x >= first:
         assert x <= maximum,  "input is above maximal contraint"
-        return (x-first)/(maximum-first)
+        return np.float(x-first) / (maximum-first)
     else:
         assert x >= minimum, "input is bellow minimum contraint"
-        return (x - first) / (first - minimum)
+        return np.float(x - first) / (first - minimum)
 
 
 def rescale(x):
